{{cookiecutter.project_slug}}/requirements/prod.txt
django==3.0.7
django-environ==0.4.5
django-imagekit==4.0.2
django-model-utils==4.0.0
django-redis==4.12.1
django-filter==2.2.0
djangorestframework==3.11.0
django-admin-tools==0.9.0
django-cors-headers==3.4.0

django-phonenumber-field==5.0.0
phonenumbers==8.12.10

celery==4.4.2
coreapi==2.3.3
psycopg2-binary==2.8.5
pillow==7.1.2
coreapi-cli==1.0.9
uwsgi==2.0.18
unidecode==1.0.23
tqdm==4.46.0

fabric==2.4.0
patchwork==1.0.1
